Initialize JavaScript variables to avoid being undefined
Created a new file understanding-uninitialized-variables.js. In this file, three JavaScript variables: a, b, c have been initialized with 5, 10, and 'I am a' respectively. This is to prevent them from having an undefined value which can result in a NaN (Not a Number) outcome or a string of undefined, during mathematical operations or string concatenations.