Add function 'timesFive' in JavaScript
A new file 'return-a-value-from-a-function-with-return.js' has been added. This file includes a function named 'timesFive' that takes a single argument, multiplies it by five and returns the result. This is to demonstrate the basis of returning a manipulated value in JavaScript functions. The function has also been tested with different arguments to ensure correctness.