Add greaterOrEqual operator functionality in JavaScript
A new file 'comparison-with-the-greater-than-or-equal-to-operator.js' has been introduced. This file features the function 'testGreaterOrEqual()' which incorporates the greater than or equal to operator in JavaScript. The function determines whether the input value is greater than or equal to 10 or 20 and returns the corresponding string message. A test case has been added for evaluation.