Add lessOrEqual operator functionality in JavaScript
Added a new file, 'comparison-with-the-less-than-or-equal-to-operator.js' which contains a function, 'testLessOrEqual'. This function tests the less than or equal to operator in JavaScript by taking a value and returning a string message that indicates if the given value is less than or equal to 12, 24 or greater than 24. Also added a test console log for verification.