The diff provided does not match the commit message. The diff shows the creation of a new file with a function that uses if-else statements to classify a number while the commit message talks about a modification that replaced 'var' with 'let'.
Here is a suitable commit message for the diff provided:

Add new file with a function using chaining if-else statements

A new file 'chaining-if-else-statements.js' was created under 'Basic JavaScript'. The file contains a function 'testSize' which uses if-else statements to return different names depending on the input number's range. This was added to demonstrate the use of multiple conditional statements in JavaScript processing flow.