Add a new function nextInLine in stand-in-line.js
A new file 'stand-in-line.js' was added under the 'Basic JavaScript' directory. In this file, a function 'nextInLine' is defined, which takes an array and a number as arguments. The function adds the number to the end of the array and removes the first element of the array and then returns it. This file and function were created to illustrate queue (FIFO) concept in JavaScript. A test setup and display codes were also added to test and verify function's operation.