Add default option in switch statement
Created a new file 'adding-a-default-option-in-switch-statements.js' in the 'Basic JavaScript' directory. The file includes the function 'switchOfStuff', which contains a switch statement that checks the input against several cases and returns a corresponding string. A default case has been added to handle any inputs that do not match the predefined cases. The addition of the default case helps to maintain the robustness of the function, ensuring it can handle a range of inputs. The function's correctness has been verified using various test cases.