Add object property testing function
This commit introduces a new JavaScript function 'checkObj'. This function takes an object and a property as arguments and checks if the object has that property. It returns the value of the property if it exists, otherwise it returns "Not Found". This function is particularly useful to avoid errors when trying to access properties that do not exist in an object.