Add iteration over odd numbers using for loop in JavaScript
Introduced a new JavaScript file 'iterate-odd-numbers-with-a-for-loop.js'. This file contains a for-loop that increments by 2 (thus iterating over odd numbers) from 1 to 9, pushing these numbers into an array. The file aims to show beginners how to use a for-loop to iterate over odd numbers and how to use array push function in JavaScript.