Add randomWholeNum function for generating random whole numbers
A new JavaScript file "generate-random-whole-numbers-with-javascript.js" was added. It contains a function, randomWholeNum(), which returns a random whole number between 0 and 10. This function leverages the Math.random() and Math.floor() methods and is useful in applications where random whole numbers are needed, enhancing the overall diversity and unpredictability in our program.