Add function for generating random whole numbers within a range
Added a new JavaScript file "generate-random-whole-numbers-within-a-range.js" that contains a function, randomRange(myMin, myMax). This function returns a random whole number in the given range. It utilizes Math.random() and Math.floor() methods. This can be useful when a random whole number within a certain range is needed in the application, thereby boosting diversity and unpredictability.