Sentences seem mixed up. The diff is about addition of a JavaScript file implementing a 'do-while' loop. Here is the corrected version:
Commit message:
Implement 'do-while' loop in JavaScript

Added a new file "iterate-with-javascript-do-while-loops.js" in the Basic JavaScript directory. The file contains a do-while loop that only pushes the number 10 to an array, so the final value of 'i' will be 11. This provides an example of how to use do-while loops in JavaScript. This is important for demonstrating a basic, yet important, concept in JavaScript programming.