Add profile-lookup functionality in JavaScript
Implemented a new file "profile-lookup.js" in the Basic JavaScript directory. It includes a function that shows how to look up for a property in an object. It takes a name and a property and checks if they exist within the given contacts array. If they do, it returns the value of that property. If not, it returns an appropriate message.